Patch to estimate alcohol consumption from HSE 2019
Updates to both the 'read_2019' and 'alc_weekmean_adult' functions in the package to accommodate survey changes in HSE 2019.

In detail for 'read_2019', specific name changes have been made to bring the variables in line with 2018 and previous years. Also some alcohol consumption variables had a positive '98.0' entry to represent 'don't know' as an answer to some self-complete questions. This has been identified as an NA string and modified accordingly.

In detail for 'alc_weekmean_adult', HSE 2019 has changed the names of many of the variables and now records both normal-strength and strong beers in pints rather than half pints for the interview. In addition, the organisation of the wine questions has changed. Small changes have also been made to the drink type preferences (from proportion to percentage) and how NA values are dealt with in sums. Finally a new variable has been created ('wine_only_units'), to preserve this data when modifying 'wine_units' to also include 'sherry_units' when aggregating drink categories.